+++ src/coordination/data_instance_management_server_handlers.cpp
@@ -16,6 +16,8 @@
 #include "coordination/include/coordination/data_instance_management_server.hpp"
 #include "replication/state.hpp"
 
+#include "rpc/utils.hpp"  // Needs to be included last so that SLK definitions are seen
+
 namespace memgraph::dbms {
 
 void DataInstanceManagementServerHandlers::Register(memgraph::coordination::DataInstanceManagementServer &server,
@@ -93,7 +95,7 @@ void DataInstanceManagementServerHandlers::StateCheckHandler(replication::Replic
   });
 
   coordination::StateCheckRes const rpc_res{is_replica, uuid, writing_enabled};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("State check returned: is_replica = {}, uuid = {}, writing_enabled = {}", is_replica,
                uuid.has_value() ? std::string{*uuid} : "", writing_enabled);
 }
@@ -102,7 +104,7 @@ void DataInstanceManagementServerHandlers::GetDatabaseHistoriesHandler(
     replication::ReplicationHandler const &replication_handler, slk::Reader * /*req_reader*/,
     slk::Builder *res_builder) {
   coordination::GetDatabaseHistoriesRes const rpc_res{replication_handler.GetDatabasesHistories()};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("Database's history returned successfully.");
 }
 
@@ -114,7 +116,7 @@ void DataInstanceManagementServerHandlers::SwapMainUUIDHandler(replication::Repl
   if (!replication_handler.IsReplica()) {
     spdlog::error("Setting uuid must be performed on replica.");
     replication_coordination_glue::SwapMainUUIDRes const rpc_res{false};
-    dbms::SendFinalResponse(rpc_res, res_builder);
+    rpc::SendFinalResponse(rpc_res, res_builder);
     return;
   }
 
@@ -124,7 +126,7 @@ void DataInstanceManagementServerHandlers::SwapMainUUIDHandler(replication::Repl
   repl_data.uuid_ = req.uuid;
 
   replication_coordination_glue::SwapMainUUIDRes const rpc_res{true};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("UUID successfully set to {}.", std::string(req.uuid));
 }
 
@@ -140,12 +142,12 @@ void DataInstanceManagementServerHandlers::DemoteMainToReplicaHandler(
   if (!replication_handler.SetReplicationRoleReplica(clients_config, std::nullopt)) {
     spdlog::error("Demoting main to replica failed.");
     coordination::DemoteMainToReplicaRes const rpc_res{false};
-    dbms::SendFinalResponse(rpc_res, res_builder);
+    rpc::SendFinalResponse(rpc_res, res_builder);
     return;
   }
 
   coordination::DemoteMainToReplicaRes const rpc_res{true};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("MAIN successfully demoted to REPLICA.");
 }
 
@@ -158,13 +160,13 @@ void DataInstanceManagementServerHandlers::GetInstanceUUIDHandler(
         "the "
         "moment. Returning empty uuid.");
     coordination::GetInstanceUUIDRes const rpc_res{{}};
-    dbms::SendFinalResponse(rpc_res, res_builder);
+    rpc::SendFinalResponse(rpc_res, res_builder);
     return;
   }
 
   auto const replica_uuid = replication_handler.GetReplicaUUID();
   coordination::GetInstanceUUIDRes const rpc_res{replica_uuid};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("Replica's UUID returned successfully: {}.", replica_uuid ? std::string{*replica_uuid} : "");
 }
 
@@ -179,7 +181,7 @@ void DataInstanceManagementServerHandlers::PromoteToMainHandler(replication::Rep
   if (const bool success = replication_handler.DoToMainPromotion(req.main_uuid); !success) {
     spdlog::error("Promoting replica to main failed.");
     coordination::PromoteToMainRes const res{false};
-    dbms::SendFinalResponse(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -196,7 +198,7 @@ void DataInstanceManagementServerHandlers::PromoteToMainHandler(replication::Rep
   replication_handler.GetReplState().GetMainRole().writing_enabled_ = true;
 
   coordination::PromoteToMainRes const res{true};
-  dbms::SendFinalResponse(res, res_builder);
+  rpc::SendFinalResponse(res, res_builder);
   spdlog::info("Promoting replica to main finished successfully. New MAIN's uuid: {}", std::string(req.main_uuid));
 }
 
@@ -205,7 +207,7 @@ void DataInstanceManagementServerHandlers::RegisterReplicaOnMainHandler(
   if (!replication_handler.IsMain()) {
     spdlog::error("Registering replica on main must be performed on main!");
     coordination::RegisterReplicaOnMainRes const res{false};
-    dbms::SendFinalResponse(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
   coordination::RegisterReplicaOnMainReq req;
@@ -218,7 +220,7 @@ void DataInstanceManagementServerHandlers::RegisterReplicaOnMainHandler(
     spdlog::error("Registering replica to main failed because MAIN's uuid {} != from coordinator's uuid {}!",
                   std::string(req.main_uuid), std::string(main_uuid));
     coordination::RegisterReplicaOnMainRes const res{false};
-    dbms::SendFinalResponse(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -229,7 +231,7 @@ void DataInstanceManagementServerHandlers::RegisterReplicaOnMainHandler(
   }
 
   coordination::RegisterReplicaOnMainRes const res{true};
-  dbms::SendFinalResponse(res, res_builder);
+  rpc::SendFinalResponse(res, res_builder);
   spdlog::info("Registering replica {} to main finished successfully.", req.replication_client_info.instance_name);
 }
 
@@ -238,7 +240,7 @@ void DataInstanceManagementServerHandlers::UnregisterReplicaHandler(
   if (!replication_handler.IsMain()) {
     spdlog::error("Unregistering replica must be performed on main.");
     coordination::UnregisterReplicaRes const res{false};
-    dbms::SendFinalResponse(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -249,31 +251,31 @@ void DataInstanceManagementServerHandlers::UnregisterReplicaHandler(
     using enum query::UnregisterReplicaResult;
     case SUCCESS: {
       coordination::UnregisterReplicaRes const rpc_res{true};
-      dbms::SendFinalResponse(rpc_res, res_builder);
+      rpc::SendFinalResponse(rpc_res, res_builder);
       break;
     }
     case NOT_MAIN: {
       spdlog::error("Unregistering replica must be performed on main.");
       coordination::UnregisterReplicaRes const rpc_res{false};
-      dbms::SendFinalResponse(rpc_res, res_builder);
+      rpc::SendFinalResponse(rpc_res, res_builder);
       break;
     }
     case CANNOT_UNREGISTER: {
       spdlog::error("Could not unregister replica.");
       coordination::UnregisterReplicaRes const rpc_res{false};
-      dbms::SendFinalResponse(rpc_res, res_builder);
+      rpc::SendFinalResponse(rpc_res, res_builder);
       break;
     }
     case COULD_NOT_BE_PERSISTED: {
       spdlog::error("Could not persist replica unregistration.");
       coordination::UnregisterReplicaRes const rpc_res{false};
-      dbms::SendFinalResponse(rpc_res, res_builder);
+      rpc::SendFinalResponse(rpc_res, res_builder);
       break;
     }
     case NO_ACCESS: {
       spdlog::error("Couldn't get unique access to ReplicationState when unregistering replica.");
       coordination::UnregisterReplicaRes const rpc_res{false};
-      dbms::SendFinalResponse(rpc_res, res_builder);
+      rpc::SendFinalResponse(rpc_res, res_builder);
       break;
     }
   }
@@ -285,19 +287,19 @@ void DataInstanceManagementServerHandlers::EnableWritingOnMainHandler(
   if (!replication_handler.IsMain()) {
     spdlog::error("Enable writing on main must be performed on main!");
     coordination::EnableWritingOnMainRes const rpc_res{false};
-    dbms::SendFinalResponse(rpc_res, res_builder);
+    rpc::SendFinalResponse(rpc_res, res_builder);
     return;
   }
 
   if (!replication_handler.GetReplState().EnableWritingOnMain()) {
     spdlog::error("Enabling writing on main failed!");
     coordination::EnableWritingOnMainRes const rpc_res{false};
-    dbms::SendFinalResponse(rpc_res, res_builder);
+    rpc::SendFinalResponse(rpc_res, res_builder);
     return;
   }
 
   coordination::EnableWritingOnMainRes const rpc_res{true};
-  dbms::SendFinalResponse(rpc_res, res_builder);
+  rpc::SendFinalResponse(rpc_res, res_builder);
   spdlog::info("Enabled writing on main.");
 }
