+++ src/coordination/include/coordination/coordinator_instance_management_server_handlers.hpp
@@ -26,14 +26,5 @@ class CoordinatorInstanceManagementServerHandlers {
                                    slk::Builder *res_builder);
 };
 
-template <typename TResponse>
-void SendFinalResponse(TResponse const &res, slk::Builder *builder) {
-  slk::Save(TResponse::kType.id, builder);
-  slk::Save(rpc::current_version, builder);
-  slk::Save(res, builder);
-  builder->Finalize();
-  spdlog::trace("[RpcServer] sent {}", TResponse::kType.name);
-}
-
 #endif
 }  // namespace memgraph::coordination
