+++ src/dbms/inmemory/replication_handlers.hpp
@@ -1,4 +1,4 @@
-// Copyright 2025 Memgraph Ltd.
+// Copyright 2024 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -58,14 +58,5 @@ class InMemoryReplicationHandlers {
   static uint64_t ReadAndApplyDeltas(storage::InMemoryStorage *storage, storage::durability::BaseDecoder *decoder,
                                      uint64_t version);
 };
-
-template <typename TResponse>
-void SendFinalResponse(TResponse const &res, slk::Builder *builder) {
-  slk::Save(TResponse::kType.id, builder);
-  slk::Save(rpc::current_version, builder);
-  slk::Save(res, builder);
-  builder->Finalize();
-  spdlog::trace("[RpcServer] sent {}", TResponse::kType.name);
-}
 
 }  // namespace memgraph::dbms
