+++ src/dbms/replication_handlers.cpp
@@ -16,6 +16,8 @@
 #include "storage/v2/storage.hpp"
 #include "system/state.hpp"
 
+#include "rpc/utils.hpp"
+
 namespace memgraph::dbms {
 
 #ifdef MG_ENTERPRISE
@@ -31,7 +33,7 @@ void CreateDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system
     spdlog::error(
         "Handling CreateDatabase, an enterprise RPC message, without license. Check your license status by running "
         "SHOW LICENSE INFO.");
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -40,7 +42,7 @@ void CreateDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system
 
   if (!current_main_uuid.has_value() || req.main_uuid != current_main_uuid) [[unlikely]] {
     LogWrongMain(current_main_uuid, req.main_uuid, memgraph::storage::replication::CreateDatabaseReq::kType.name);
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -52,7 +54,7 @@ void CreateDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system
   if (req.expected_group_timestamp != system_state_access.LastCommitedTS()) {
     spdlog::debug("CreateDatabaseHandler: bad expected timestamp {},{}", req.expected_group_timestamp,
                   system_state_access.LastCommitedTS());
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -69,7 +71,7 @@ void CreateDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system
     // Failure
   }
 
-  memgraph::slk::Save(res, res_builder);
+  rpc::SendFinalResponse(res, res_builder);
 }
 
 void DropDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system_state_access,
@@ -83,7 +85,7 @@ void DropDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system_s
     spdlog::error(
         "Handling DropDatabase, an enterprise RPC message, without license. Check your license status by running SHOW "
         "LICENSE INFO.");
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -92,7 +94,7 @@ void DropDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system_s
 
   if (!current_main_uuid.has_value() || req.main_uuid != current_main_uuid) [[unlikely]] {
     LogWrongMain(current_main_uuid, req.main_uuid, memgraph::storage::replication::DropDatabaseReq::kType.name);
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -104,7 +106,7 @@ void DropDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system_s
   if (req.expected_group_timestamp != system_state_access.LastCommitedTS()) {
     spdlog::debug("DropDatabaseHandler: bad expected timestamp {},{}", req.expected_group_timestamp,
                   system_state_access.LastCommitedTS());
-    memgraph::slk::Save(res, res_builder);
+    rpc::SendFinalResponse(res, res_builder);
     return;
   }
 
@@ -127,7 +129,7 @@ void DropDatabaseHandler(memgraph::system::ReplicaHandlerAccessToState &system_s
     // Failure
   }
 
-  memgraph::slk::Save(res, res_builder);
+  rpc::SendFinalResponse(res, res_builder);
 }
 
 bool SystemRecoveryHandler(DbmsHandler &dbms_handler, const std::vector<storage::SalientConfig> &database_configs) {
