+++ src/replication_coordination_glue/handler.hpp
@@ -15,17 +15,9 @@
 
 #include "messages.hpp"
 #include "rpc/messages.hpp"
+#include "rpc/utils.hpp"  // Needs to be included last so that SLK definitions are seen
 
 namespace memgraph::replication_coordination_glue {
-
-template <typename TResponse>
-void SendFinalResponse(TResponse const &res, slk::Builder *builder) {
-  slk::Save(TResponse::kType.id, builder);
-  slk::Save(rpc::current_version, builder);
-  slk::Save(res, builder);
-  builder->Finalize();
-  spdlog::trace("[RpcServer] sent {}", TResponse::kType.name);
-}
 
 inline bool SendSwapMainUUIDRpc(memgraph::rpc::Client &rpc_client_, const memgraph::utils::UUID &uuid) {
   try {
@@ -46,6 +38,6 @@ inline void FrequentHeartbeatHandler(slk::Reader *req_reader, slk::Builder *res_
   FrequentHeartbeatReq::Load(&req, req_reader);
   memgraph::slk::Load(&req, req_reader);
   FrequentHeartbeatRes res{};
-  SendFinalResponse(res, res_builder);
+  rpc::SendFinalResponse(res, res_builder);
 }
 }  // namespace memgraph::replication_coordination_glue
