+++ tests/unit/rpc_timeouts.cpp
@@ -12,10 +12,11 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
+#include "rpc_messages.hpp"
+
 #include "rpc/client.hpp"
 #include "rpc/server.hpp"
-
-#include "rpc_messages.hpp"
+#include "rpc/utils.hpp"  // Needs to be included last so that SLK definitions are seen
 
 using memgraph::communication::ClientContext;
 using memgraph::communication::ServerContext;
@@ -60,14 +61,6 @@ void EchoMessage::Save(const EchoMessage &obj, memgraph::slk::Builder *builder) 
 
 constexpr int port{8181};
 
-template <typename TResponse>
-void SendFinalResponse(TResponse const &res, memgraph::slk::Builder *builder) {
-  memgraph::slk::Save(TResponse::kType.id, builder);
-  memgraph::slk::Save(memgraph::rpc::current_version, builder);
-  memgraph::slk::Save(res, builder);
-  builder->Finalize();
-}
-
 // RPC client is setup with timeout but shouldn't be triggered.
 TEST(RpcTimeout, TimeoutNoFailure) {
   Endpoint endpoint{"localhost", port};
@@ -84,7 +77,7 @@ TEST(RpcTimeout, TimeoutNoFailure) {
     Load(&req, req_reader);
 
     EchoMessage res{"Sending reply"};
-    SendFinalResponse(res, res_builder);
+    memgraph::rpc::SendFinalResponse(res, res_builder);
   });
 
   ASSERT_TRUE(rpc_server.Start());
@@ -116,7 +109,7 @@ TEST(RpcTimeout, TimeoutExecutionBlocks) {
 
     std::this_thread::sleep_for(1100ms);
     EchoMessage res{"Sending reply"};
-    SendFinalResponse(res, res_builder);
+    memgraph::rpc::SendFinalResponse(res, res_builder);
   });
 
   ASSERT_TRUE(rpc_server.Start());
@@ -147,7 +140,7 @@ TEST(RpcTimeout, TimeoutServerBusy) {
     Load(&req, req_reader);
     std::this_thread::sleep_for(2500ms);
     SumRes res(req.x + req.y);
-    SendFinalResponse(res, res_builder);
+    memgraph::rpc::SendFinalResponse(res, res_builder);
   });
 
   rpc_server.Register<Echo>([](auto *req_reader, auto *res_builder) {
@@ -156,7 +149,7 @@ TEST(RpcTimeout, TimeoutServerBusy) {
     Load(&req, req_reader);
 
     EchoMessage res{"Sending reply"};
-    SendFinalResponse(res, res_builder);
+    memgraph::rpc::SendFinalResponse(res, res_builder);
   });
 
   ASSERT_TRUE(rpc_server.Start());
@@ -196,7 +189,7 @@ TEST(RpcTimeout, SendingToWrongSocket) {
 
     std::this_thread::sleep_for(1100ms);
     EchoMessage res{"Sending reply"};
-    SendFinalResponse(res, res_builder);
+    memgraph::rpc::SendFinalResponse(res, res_builder);
   });
 
   ASSERT_TRUE(rpc_server.Start());
